+++ Assets/Mirror/Examples/Lobby/Scripts/PlayerController.cs
@@ -18,9 +18,6 @@ namespace Mirror.Examples.NetworkLobby
         [SyncVar]
         public uint score;
 
-        [SyncVar(hook = nameof(SetColor))]
-        public Color playerColor = Color.black;
-
         public override void OnStartLocalPlayer()
         {
             base.OnStartLocalPlayer();
@@ -31,9 +28,22 @@ namespace Mirror.Examples.NetworkLobby
             Camera.main.enabled = false;
         }
 
+        [SyncVar(hook = nameof(SetColor))]
+        public Color playerColor = Color.black;
+
+        // Unity makes a clone of the material when GetComponent<Renderer>().material is used
+        // Cache it here and Destroy it in OnDestroy to prevent a memory leak
+        Material materialClone;
+
         void SetColor(Color color)
         {
-            GetComponent<Renderer>().material.color = color;
+            if (materialClone == null) materialClone = GetComponent<Renderer>().material;
+            materialClone.color = color;
+        }
+
+        private void OnDestroy()
+        {
+            Destroy(materialClone);
         }
 
         float horizontal = 0f;
@@ -100,7 +110,10 @@ namespace Mirror.Examples.NetworkLobby
         public void CmdClaimPrize(GameObject hitObject)
         {
             // Null check is required, otherwise close timing of multiple claims could throw a null ref.
-            hitObject?.GetComponent<Reward>().ClaimPrize(gameObject);
+            if (hitObject != null)
+            {
+                hitObject.GetComponent<Reward>().ClaimPrize(gameObject);
+            }
         }
 
         void OnGUI()